challenge-2/requirements.txt
numpy
confluent-kafka
oauth2client
google-cloud-bigquery